+++ matrix-sdk/src/main/java/org/matrix/androidsdk/rest/model/WellKnownBaseConfig.kt
@@ -27,6 +27,7 @@ import com.google.gson.annotations.SerializedName
  */
 class WellKnownBaseConfig {
 
+    @JvmField
     @SerializedName("base_url")
     var baseURL: String? = null
 }